# Conflicts:
#	AdminInterface/application/views/templates/header.php
#	AdminInterface/application/views/templates/sidebar.php
#	AdminInterface/public/css/style.css

+++ AdminInterface/application/controllers/Delete.php
@@ -13,46 +13,42 @@ class Delete extends CI_Controller {
         }
     }
 
-    public function delete_school($school_id)
-    {
+    public function delete_school($school_id) {
         $this->database_model->delete_school($school_id);
         redirect(base_url("Koolid"));
     }
 
-    public function delete_class($class_id)
-    {
+    public function delete_class($class_id) {
         $this->database_model->delete_class($class_id);
         redirect(base_url("Klassid"));
     }
 
-    public function delete_book($book_id)
-    {
+    public function delete_book($book_id) {
         $this->database_model->delete_book($book_id);
         redirect(base_url("Raamatud"));
     }
 
-    public function delete_from_list($id)
-    {
+    public function delete_from_list($id) {
         $list_item = $this->database_model->get_from_reading_list($id);
         $class_id = $list_item['class_id'];
         $this->database_model->delete_from_reading_list($id);
         redirect(base_url("Muuda/Nimekiri/".$class_id));
     }
 
-    public function delete_list($id)
-    {
+    public function delete_list($id) {
         $this->database_model->delete_reading_list($id);
         redirect(base_url("Nimekiri"));
     }
 
-    public function delete_user($user_id)
-    {
+    public function delete_user($user_id) {
+        if ($_SESSION['is_admin'] != 1) {
+            redirect(base_url('Koolid'));
+        }
         $this->database_model->delete_user($user_id);
         redirect(base_url("Kasutajad"));
     }
 
-    public function delete_keyword($keyword_id)
-    {
+    public function delete_keyword($keyword_id) {
         $book_keyword_entries = $this->database_model->get_keywords(NULL, $keyword_id);
         foreach ($book_keyword_entries as $entry) {
             $this->database_model->delete_keyword_from_book($entry['id']);
@@ -61,15 +57,13 @@ class Delete extends CI_Controller {
         redirect(base_url("Märksõnad"));
     }
 
-    public function delete_keyword_from_book($id)
-    {
+    public function delete_keyword_from_book($id) {
         $book_id = $this->database_model->get_book_keyword_entry($id)['book_id'];
         $this->database_model->delete_keyword_from_book($id);
         redirect(base_url("Muuda/Raamat/".$book_id));
     }
 
-    public function delete_author($author_id)
-    {
+    public function delete_author($author_id) {
         $book_keyword_entries = $this->database_model->get_authors(NULL, $author_id);
         foreach ($book_keyword_entries as $entry) {
             $this->database_model->delete_author_from_book($entry['id']);
@@ -78,15 +72,13 @@ class Delete extends CI_Controller {
         redirect(base_url("Autorid"));
     }
 
-    public function delete_author_from_book($id)
-    {
+    public function delete_author_from_book($id) {
         $book_id = $this->database_model->get_book_author_entry($id)['book_id'];
         $this->database_model->delete_author_from_book($id);
         redirect(base_url("Muuda/Raamat/".$book_id));
     }
 
-    public function delete_genre($genre_id)
-    {
+    public function delete_genre($genre_id) {
         $book_genre_entries = $this->database_model->get_genres(NULL, $genre_id);
         foreach ($book_genre_entries as $entry) {
             $this->database_model->delete_genre_from_book($entry['id']);
@@ -95,8 +87,7 @@ class Delete extends CI_Controller {
         redirect(base_url("Zanrid"));
     }
 
-    public function delete_genre_from_book($id)
-    {
+    public function delete_genre_from_book($id) {
         $book_id = $this->database_model->get_book_genre_entry($id)['book_id'];
         $this->database_model->delete_genre_from_book($id);
         redirect(base_url("Muuda/Raamat/".$book_id));
